moving flash_errors to remedyutils
